Move the float fix to doit() to cover related issues
