+++ sympy/integrals/risch.py
@@ -29,7 +29,7 @@ from functools import reduce
 from sympy.core.function import Lambda
 from sympy.core.mul import Mul
 from sympy.core.intfunc import ilcm
-from sympy.core.numbers import I, Rational, Float
+from sympy.core.numbers import I
 from sympy.core.power import Pow
 from sympy.core.relational import Ne
 from sympy.core.singleton import S
@@ -1702,20 +1702,6 @@ class NonElementaryIntegral(Integral):
     pass
 
 
-def _replace_coeff(e, q, r):
-    if e.is_Pow:
-        return Pow(_replace_coeff(e.base, q, r), e.exp)
-    if isinstance(e, exp):
-        return e
-    if q(e):
-        return r(e)
-    args = getattr(e, 'args', ())
-    fargs = tuple(_replace_coeff(a, q, r) for a in args)
-    if args != fargs:
-        return e.func(*fargs)
-    return e
-
-
 def risch_integrate(f, x, extension=None, handle_first='log',
                     separate_integral=False, rewrite_complex=None,
                     conds='piecewise'):
@@ -1824,9 +1810,7 @@ def risch_integrate(f, x, extension=None, handle_first='log',
     log(log(x))
 
     """
-    f0 = S(f)
-    precision = max([0] + [n._prec for n in f0.atoms(Float)])
-    f = _replace_coeff(f0, lambda e: e.is_Float, Rational)
+    f = S(f)
 
     DE = extension or DifferentialExtension(f, x, handle_first=handle_first,
             dummy=True, rewrite_complex=rewrite_complex)
@@ -1853,11 +1837,6 @@ def risch_integrate(f, x, extension=None, handle_first='log',
         else:
             raise NotImplementedError("Only exponential and logarithmic "
             "extensions are currently supported.")
-
-        if f != f0:
-            is_rational = lambda e: e.is_Rational or e.is_Integer
-            as_float = lambda n: Float(n, precision=precision)
-            ans = _replace_coeff(ans, is_rational, as_float)
 
         result += ans
         if b:
